fixed the filters but not the tests and chips are missing :(
