+++ package.json
@@ -115,8 +115,8 @@
     "@babel/preset-env": "7.15.4",
     "@babel/preset-flow": "7.14.5",
     "@babel/preset-react": "7.14.5",
-    "@cypress/code-coverage": "^3.10.0",
-    "@cypress/webpack-dev-server": "^2.0.0",
+    "@cypress/code-coverage": "^3.12.36",
+    "@cypress/webpack-dev-server": "^3.8.0",
     "@formatjs/cli": "4.2.33",
     "@redhat-cloud-services/eslint-config-redhat-cloud-services": "1.2.1",
     "@redhat-cloud-services/frontend-components-config": "^5.1.1",
@@ -132,7 +132,7 @@
     "babel-plugin-lodash": "3.3.4",
     "commander": "8.1.0",
     "copy-webpack-plugin": "9.0.1",
-    "cypress": "^12.13.0",
+    "cypress": "^13.8.1",
     "eslint": "7.32.0",
     "eslint-loader": "4.0.2",
     "eslint-plugin-cypress": "^2.12.1",
